requirements.txt
# GraphGuard — Python Dependencies
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
google-genai>=1.0.0
chromadb>=0.4.0
PyYAML>=6.0
